KVM: Move kvm_vcpu fields into common structure
This commit is in preparation for supporting multiple VM privilege
levels (VMPLs) within KVM by creating an instance of struct kvm_vcpu
for each VMPL. Support for multiple VMPLs will be introduced in a later
commit. This commit prepares struct kvm_vcpu by allowing the fields that
are common to all VMPLs to be accessed from the struct kvm_vcpu for any
VMPL.

To add support for common fields across multiple struct kvm_vcpus, this
commit creates a sub-structure within struct kvm_vcpu for storing the
common fields. Any fields that have been identified as common across
VMPLs for SEV-SNP (for now) have been moved into this sub-structure.

When the struct kvm_vcpu is initialised, a pointer to the common
structure is stored within struct kvm_vcpu allowing access to the common
fields.

This means that any reference to these common fields needs to be updated
to access them via the structure pointer, e.g.
vcpu->common->[common_field]. This has resulted in this patch generating
many small changes to many files which is an unfortunate side-effect of
this implementation.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>
[LL] Code adjusted with some changes introduced in 6.12
Signed-off-by: Luigi Leonardi <leonardi@redhat.com>