vsock/test: factor out vsock_accept from vsock_listen_accept
Remove accept from `vsock_listen_accept` and move it to the new function
`vsock_accept`. This is to increase code reusability.

This new function checks for errors and if it does not fail the caller
can be sure that `accept` has succeeded. For this reason drop all checks
on fd after every function that uses it.

Signed-off-by: Luigi Leonardi <leonardi@redhat.com>